Move cypress integration files out to root
