+++ integrations/cypress/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../../tsconfig-base.json",
   "compilerOptions": {
+    "types": ["cypress"],
     "outDir": "dist/cjs",
     "rootDir": "./src",
     "declarationDir": "dist"
