+++ packages/interactor/src/index.ts
@@ -16,5 +16,3 @@ export { CheckBox } from './definitions/check-box';
 export { RadioButton } from './definitions/radio-button';
 export { Select } from './definitions/select';
 export { MultiSelect } from './definitions/multi-select';
-
-import './integrations/cypress';
